+++ mcp_server/engine/policies/__init__.py
@@ -34,6 +34,7 @@ from mcp_server.engine.policies.anti_regression import AntiRegression
 from mcp_server.engine.policies.blast_radius import BlastRadiusVeto
 from mcp_server.engine.policies.decision_lock import DecisionLock
 from mcp_server.engine.policies.relevance_inject import RelevanceInject
+from mcp_server.engine.policies.session_log_enforcer import SessionLogEnforcer
 from mcp_server.engine.policies.token_budget import TokenBudgetPersist
 
 __all__ = [
@@ -41,5 +42,6 @@ __all__ = [
     "BlastRadiusVeto",
     "DecisionLock",
     "RelevanceInject",
+    "SessionLogEnforcer",
     "TokenBudgetPersist",
 ]